requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
fastavro==1.8.3
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
protobuf==4.24.3
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7
xmltodict==0.13.0
